@@ -1,6 +1,6 @@
 [metadata]
 name = yt_idv
-version = 0.2.3
+version = 0.3.0
 description = Interactive Volume Rendering for yt
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -8,7 +8,7 @@ url = https://github.com/yt-project/yt_idv
 author = Matthew Turk
 author_email = matthewturk@gmail.com
 license = BSD-3-Clause
-license_file = LICENSE
+license_files = LICENSE
 classifiers =
     Development Status :: 2 - Pre-Alpha
     Intended Audience :: Developers
@@ -16,10 +16,10 @@ classifiers =
     Natural Language :: English
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Programming Language :: Python :: Implementation :: CPython
 keywords = yt_idv
 
@@ -35,8 +35,8 @@ install_requires =
     pyyaml>=5.3.1
     traitlets>=5.0.5
     traittypes>=0.2.1
-    yt>=4.0.0
-python_requires = >=3.6
+    yt>=4.1.0
+python_requires = >=3.7
 include_package_data = True
 zip_safe = False
 
@@ -72,15 +72,16 @@ universal = 1
 max-line-length = 88
 exclude = docs,
     */__init__.py
-ignore = E203, # Whitespace before ':' (black compatibility)
-    E231, # Missing whitespace after ',', ';', or ':'
-    E266, # Too many leading '#' for block comment
-    E302, # Expected 2 blank lines, found 0
-    E306, # Expected 1 blank line before a nested definition
-    E741, # Do not use variables named 'I', 'O', or 'l'
-    W503, # Line break occurred before a binary operator (black compatibility)
-    W605, # Invalid escape sequence 'x'
-    B302, # this is a python 3 compatibility warning, not relevant since don't support python 2 anymore
+ignore =
+    E203,
+    E231,
+    E266,
+    E302,
+    E306,
+    E741,
+    W503,
+    W605,
+    B302,
 
 [aliases]
 test = pytest
