@@ -8,7 +8,8 @@ from .base_context import BaseContext
 
 
 class PygletRenderingContext(pyglet.window.Window, BaseContext):
-    """Basic rendering context for IDV using GLFW3, that handles the main window even loop
+    """
+    Basic rendering context for IDV using GLFW3, that handles the main window event loop
 
     Parameters
     ----------
@@ -47,8 +48,8 @@ class PygletRenderingContext(pyglet.window.Window, BaseContext):
             double_buffer=True,
             depth_size=24,
         )
-        super(PygletRenderingContext, self).__init__(
-            width, height, config=config, visible=visible, caption=title
+        super().__init__(
+            width, height, config=config, visible=visible, caption=title, resizable=True
         )
         if position is None:
             self.center_window()
@@ -77,6 +78,13 @@ class PygletRenderingContext(pyglet.window.Window, BaseContext):
         if self.gui:
             self.switch_to()
             self.gui.render(self.scene)
+
+    def on_resize(self, width, height):
+        super().on_resize(width, height)
+        self.scene.reset_framebuffers()
+        self.scene.camera.aspect_ratio = width / height
+        self.scene.camera._update_matrices()
+        self._do_update = True
 
     def set_position(self, xpos, ypos):
         if xpos < 0 or ypos < 0:
