@@ -4,6 +4,7 @@ import numpy as np
 import traitlets
 from OpenGL import GL
 
+from yt_idv.gui_support import add_popup_help
 from yt_idv.opengl_support import TransferFunctionTexture
 from yt_idv.scene_components.base_component import SceneComponent
 from yt_idv.scene_data.block_collection import BlockCollection
@@ -26,11 +27,13 @@ class BlockRendering(SceneComponent):
     tf_min = traitlets.CFloat(0.0)
     tf_max = traitlets.CFloat(1.0)
     tf_log = traitlets.Bool(True)
+    slice_position = traitlets.Tuple((0.5, 0.5, 0.5), trait=traitlets.CFloat())
+    slice_normal = traitlets.Tuple((1.0, 0.0, 0.0), trait=traitlets.CFloat())
 
     priority = 10
 
     def render_gui(self, imgui, renderer, scene):
-        changed = super(BlockRendering, self).render_gui(imgui, renderer, scene)
+        changed = super().render_gui(imgui, renderer, scene)
         _, sample_factor = imgui.slider_float(
             "Sample Factor", self.sample_factor, 1.0, 20.0
         )
@@ -103,6 +106,21 @@ class BlockRendering(SceneComponent):
                     data[xb1:xb2, 0, i] = np.mgrid[yv1 : yv2 : (xb2 - xb1) * 1j]
             if update:
                 self.transfer_function.data = (data * 255).astype("u1")
+
+        elif self.render_method == "slice":
+            imgui.text("Set slicing parameters:")
+
+            _, self.slice_position = imgui.input_float3(
+                "Position", *self.slice_position
+            )
+            changed = changed or _
+            _ = add_popup_help(imgui, "The position of a point on the slicing plane.")
+            changed = changed or _
+            _, self.slice_normal = imgui.input_float3("Normal", *self.slice_normal)
+            changed = changed or _
+            _ = add_popup_help(imgui, "The normal vector of the slicing plane.")
+            changed = changed or _
+
         return changed
 
     @traitlets.default("transfer_function")
@@ -123,9 +141,27 @@ class BlockRendering(SceneComponent):
     def _set_uniforms(self, scene, shader_program):
         shader_program._set_uniform("box_width", self.box_width)
         shader_program._set_uniform("sample_factor", self.sample_factor)
-        shader_program._set_uniform("ds_tex", 0)
+        shader_program._set_uniform("ds_tex", np.array([0, 0, 0, 0, 0, 0]))
         shader_program._set_uniform("bitmap_tex", 1)
         shader_program._set_uniform("tf_tex", 2)
         shader_program._set_uniform("tf_min", self.tf_min)
         shader_program._set_uniform("tf_max", self.tf_max)
         shader_program._set_uniform("tf_log", float(self.tf_log))
+        shader_program._set_uniform("slice_normal", np.array(self.slice_normal))
+        shader_program._set_uniform("slice_position", np.array(self.slice_position))
+
+    def _get_sanitized_iso_layers(self):
+        # return the sanitized layers
+
+        # pull extrema from `BlockCollection`, these are extrema across all blocks
+        data_min = self.data.min_val
+        data_max = self.data.max_val
+        if self.cmap_log:
+            data_min = np.log10(data_min)
+            data_max = np.log10(data_max)
+
+        # apply the same scaling as in `self.data._load_textures()`
+        normalized_isovals = [
+            (iso_val - data_min) / (data_max - data_min) for iso_val in self.iso_layers
+        ]
+        return normalized_isovals
